Refactor server file paths for calculator examples
