+++ examples/calculator_example/framed_pipe/client.cpp
