+++ examples/calculator_example/framed_socket/client.cpp
