+++ examples/calculator_example/pipe/client.cpp
