+++ examples/calculator_example/pipe/server.cpp
@@ -1,16 +1,15 @@
 #include <memory>
 
 #include <jsonrpc/server/server.hpp>
-#include <jsonrpc/transport/framed_pipe_transport.hpp>
+#include <jsonrpc/transport/pipe_transport.hpp>
 #include <nlohmann/json.hpp>
 #include <spdlog/sinks/basic_file_sink.h>
 #include <spdlog/spdlog.h>
 
-#include "calculator.hpp"
-#include "jsonrpc/transport/framed_socket_transport.hpp"
+#include "../calculator.hpp"
 
 using jsonrpc::server::Server;
-using jsonrpc::transport::FramedSocketTransport;
+using jsonrpc::transport::PipeTransport;
 using Json = nlohmann::json;
 
 auto main() -> int {
@@ -19,9 +18,9 @@ auto main() -> int {
   spdlog::set_level(spdlog::level::debug);
   spdlog::flush_on(spdlog::level::debug);
 
-  const std::string host = "0.0.0.0";
-  const uint16_t port = 12345;
-  auto transport = std::make_unique<FramedSocketTransport>(host, port, true);
+  const std::string socket_path = "/tmp/calculator_pipe";
+
+  auto transport = std::make_unique<PipeTransport>(socket_path, true);
   Server server(std::move(transport));
 
   server.RegisterMethodCall("add", [](const std::optional<Json> &params) {
