+++ examples/calculator_example/socket/client.cpp
