+++ examples/calculator_example/stdio/client.cpp
